requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Login==0.6.3

psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

PyJWT==2.8.0
requests==2.31.0

python-dotenv==1.0.0

Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3 